# group76_part1/api-spec.yaml
openapi: 3.0.3

info:
  title: SafeDispatch HR Administration API
  description: |
    REST API for SafeDispatch office administrators to manage employees,
    departments, assignments, and employee promotions.
  version: 1.0.0

servers:
  - url: https://localhost:8443
    description: Local HTTPS server

tags:
  - name: Employees
    description: Employee management operations
  - name: Departments
    description: Department management operations
  - name: Assignments
    description: Employee-department assignment operations
  - name: Promotions
    description: Promotion business logic operations

security:
  - basicAuth: []

paths:
  /api/employees:
    get:
      tags: [Employees]
      summary: List all employees
      description: Returns all employees. Can be filtered by department name.
      parameters:
        - in: query
          name: department
          required: false
          schema:
            type: string
          description: Department name to filter by
          example: Logistics
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeResponse'
              examples:
                success:
                  value:
                    - id: 1
                      name: Alice Johnson
                      contractType: FULL_TIME
                      startDate: "2023-10-01"
                      salary: 33000.0
                      email: alice@safedispatch.com
                      address: Liverpool
                      rank: MID
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    post:
      tags: [Employees]
      summary: Create a new employee
      requestBody:
        $ref: '#/components/requestBodies/EmployeeRequestBody'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
              examples:
                created:
                  value:
                    id: 1
                    name: Alice Johnson
                    contractType: FULL_TIME
                    startDate: "2023-10-01"
                    salary: 30000.0
                    email: alice@safedispatch.com
                    address: Liverpool
                    rank: JUNIOR
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

  /api/employees/{id}:
    get:
      tags: [Employees]
      summary: Get employee by ID
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '200':
          description: Employee retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
              examples:
                success:
                  value:
                    id: 1
                    name: Alice Johnson
                    contractType: FULL_TIME
                    startDate: "2023-10-01"
                    salary: 33000.0
                    email: alice@safedispatch.com
                    address: Liverpool
                    rank: MID
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    put:
      tags: [Employees]
      summary: Update employee details
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        $ref: '#/components/requestBodies/EmployeeRequestBody'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    delete:
      tags: [Employees]
      summary: Delete an employee
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '204':
          description: Employee deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

  /api/employees/{id}/promote:
    put:
      tags: [Promotions]
      summary: Promote an employee
      description: |
        Promotes an employee if they have worked for at least 6 months.
        This operation updates both rank and salary.
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '200':
          description: Employee promoted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionResponse'
              examples:
                promoted:
                  value:
                    employeeId: 1
                    employeeName: Alice Johnson
                    oldRank: JUNIOR
                    newRank: MID
                    oldSalary: 30000.0
                    newSalary: 33000.0
                    increaseAmount: 3000.0
                    message: Employee promoted successfully
        '400':
          description: Employee not eligible for promotion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                notEligible:
                  value:
                    error: Employee is not eligible for promotion yet
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

  /api/departments:
    get:
      tags: [Departments]
      summary: List all departments
      responses:
        '200':
          description: Departments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentResponse'
              examples:
                success:
                  value:
                    - id: 1
                      name: Logistics
                      budget: 100000.0
                      location: Liverpool
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    post:
      tags: [Departments]
      summary: Create a department
      requestBody:
        $ref: '#/components/requestBodies/DepartmentRequestBody'
      responses:
        '201':
          description: Department created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
              examples:
                created:
                  value:
                    id: 1
                    name: Logistics
                    budget: 100000.0
                    location: Liverpool
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

  /api/assignments:
    post:
      tags: [Assignments]
      summary: Assign an employee to a department
      requestBody:
        $ref: '#/components/requestBodies/AssignmentRequestBody'
      responses:
        '201':
          description: Assignment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
              examples:
                created:
                  value:
                    id: 1
                    employeeId: 1
                    employeeName: Alice Johnson
                    departmentId: 1
                    departmentName: Logistics
                    role: Member
                    accessLevel: Standard
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

  /api/assignments/{id}:
    delete:
      tags: [Assignments]
      summary: Delete an assignment
      parameters:
        - in: path
          name: id
          required: true
          description: Assignment ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Assignment deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  parameters:
    EmployeeId:
      in: path
      name: id
      required: true
      description: Employee ID
      schema:
        type: integer
        format: int64

  requestBodies:
    EmployeeRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmployeeRequest'
          examples:
            employee:
              value:
                name: Alice Johnson
                contractType: FULL_TIME
                startDate: "2023-10-01"
                salary: 30000.0
                email: alice@safedispatch.com
                address: Liverpool
                rank: JUNIOR

    DepartmentRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DepartmentRequest'
          examples:
            department:
              value:
                name: Logistics
                budget: 100000.0
                location: Liverpool

    AssignmentRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AssignmentRequest'
          examples:
            assignment:
              value:
                employeeId: 1
                departmentId: 1
                role: Member
                accessLevel: Standard

  responses:
    BadRequestResponse:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            badRequest:
              value:
                error: Validation failed

    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            notFound:
              value:
                error: Employee not found with id 99

    UnauthorizedResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorized:
              value:
                error: Unauthorized

  schemas:
    EmployeeRequest:
      type: object
      required:
        - name
        - contractType
        - startDate
        - salary
        - email
        - address
      properties:
        name:
          type: string
          example: Alice Johnson
        contractType:
          type: string
          example: FULL_TIME
        startDate:
          type: string
          format: date
          example: "2023-10-01"
        salary:
          type: number
          format: double
          example: 30000.0
        email:
          type: string
          format: email
          example: alice@safedispatch.com
        address:
          type: string
          example: Liverpool
        rank:
          type: string
          example: JUNIOR

    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contractType:
          type: string
        startDate:
          type: string
          format: date
        salary:
          type: number
          format: double
        email:
          type: string
        address:
          type: string
        rank:
          type: string

    DepartmentRequest:
      type: object
      required:
        - name
        - budget
        - location
      properties:
        name:
          type: string
          example: Logistics
        budget:
          type: number
          format: double
          example: 100000.0
        location:
          type: string
          example: Liverpool

    DepartmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        budget:
          type: number
          format: double
        location:
          type: string

    AssignmentRequest:
      type: object
      required:
        - employeeId
        - departmentId
      properties:
        employeeId:
          type: integer
          format: int64
          example: 1
        departmentId:
          type: integer
          format: int64
          example: 1
        role:
          type: string
          example: Member
        accessLevel:
          type: string
          example: Standard

    AssignmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        departmentId:
          type: integer
          format: int64
        departmentName:
          type: string
        role:
          type: string
        accessLevel:
          type: string

    PromotionResponse:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        oldRank:
          type: string
        newRank:
          type: string
        oldSalary:
          type: number
          format: double
        newSalary:
          type: number
          format: double
        increaseAmount:
          type: number
          format: double
        message:
          type: string

    ApiError:
      type: object
      properties:
        error:
          type: string
          example: Employee is not eligible for promotion yet